Moved abstract basic user into basic user module.
