--- core/AbstractModels/AbstractBaseUser/apps.py
@@ -1,9 +1,0 @@
-"""App definition for Abstract User."""
-from django.apps import AppConfig
-
-
-class AbstractbaseuserConfig(AppConfig):
-    """App definition for Abstract User."""
-
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'AbstractModels.AbstractBaseUser'
