--- core/AbstractModels/README.md
@@ -1,5 +1,0 @@
-# Abstract Models
-This section contains all abstract models.
-
-## Folder Structure
-- AbstractBaseUser/ A basic abstracted user used for custom inheritance.
