+++ core/DjangoPlusPlus/settings.py
@@ -26,7 +26,6 @@ INSTALLED_APPS = [
     'rest_framework',
     'django_rest_passwordreset',
     'knox',
-    'AbstractModels.AbstractBaseUser',
     'BasicUser',
 ]
