+++ core/README.md
@@ -2,6 +2,5 @@
 Contains the core code of Django++.
 
 ## Folder Structure
-- AbstractModels/ Contains all abstract models.
 - BasicUser/ Contains a module that represent common users, is the projects default AUTH_USER_MODEL.
 - TestUtil/ Contains various testing tools.
